# cc-registry-v2/backend/openapi.yaml
openapi: 3.1.0
info:
  title: RunWhen CodeCollection Registry API
  description: |
    Backend API for the RunWhen CodeCollection Registry. Manages codecollections,
    codebundles, tasks, documentation, AI enhancement, Helm charts, and chat.

    ## Architecture

    This API is the single source of truth for all registry data. It is consumed by:
    - **Frontend** — React UI at the same origin
    - **MCP Server** — stateless tool server for LLM clients (via `REGISTRY_API_URL`)
    - **Celery Workers** — background task processing (sync, parse, enhance, index)

    ## Authentication

    Admin endpoints require a Bearer token in the `Authorization` header.
    Public endpoints (registry, chat, analytics) require no authentication.

    ## Data Flow

    ```
    codecollections.yaml → Seed → Sync (git clone) → Parse (codebundles) → Enhance (AI) → Serve
    ```

  version: 2.0.0
  contact:
    name: RunWhen
    url: https://runwhen.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8001
    description: Local development
  - url: http://cc-registry-backend:8001
    description: In-cluster (Kubernetes)
...
paths:
...
  /api/v1/registry/collections/{collection_slug}:
    get:
      tags: [Registry]
      summary: Get collection by slug
      operationId: getCollection
      parameters:
        - $ref: "#/components/parameters/CollectionSlug"
      responses:
        "200":
          description: Collection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionWithStats"
        "404":
          description: Collection not found
  /api/v1/codebundles:
    get:
      tags: [Registry]
      summary: List/search codebundles
      description: |
        List codebundles with full-text search, filtering, and pagination.
        The `search` parameter searches across name, display_name, description, and doc fields.
      operationId: listCodebundles
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: search
          in: query
          schema: { type: string }
          description: Full-text search across name, display_name, description, doc
        - name: collection_id
          in: query
          schema: { type: integer }
        - name: tags
          in: query
          schema: { type: string }
          description: Comma-separated support tags
        - name: platform
          in: query
          schema: { type: string }
          description: Platform filter (e.g. AWS, Kubernetes, Azure)
        - name: access_level
          in: query
          schema: { type: string, enum: [read-only, read-write, unknown] }
        - name: has_auto_discovery
          in: query
          schema: { type: boolean }
        - name: sort_by
          in: query
          schema: { type: string, enum: [name, updated, tasks], default: name }
      responses:
        "200":
          description: Paginated codebundle list
          content:
            application/json:
              schema:
                type: object
                properties:
                  codebundles:
                    type: array
                    items:
                      $ref: "#/components/schemas/CodeBundle"
                  total_count: { type: integer }
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /api/v1/collections/{collection_slug}/codebundles/{codebundle_slug}:
    get:
      tags: [Registry]
      summary: Get codebundle by slug
      operationId: getCodebundle
      parameters:
        - $ref: "#/components/parameters/CollectionSlug"
        - name: codebundle_slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full codebundle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeBundleFull"
        "404":
          description: Codebundle not found
  /api/v1/registry/tasks:
    get:
      tags: [Registry]
      summary: Search tasks
      description: Search tasks across all codebundles with tag and text filtering.
      operationId: searchTasks
      parameters:
        - name: support_tags
          in: query
          schema: { type: string }
          description: Comma-separated tags
        - name: search
          in: query
          schema: { type: string }
        - name: collection_slug
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Task search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks: { type: array, items: { type: object } }
                  total_count: { type: integer }
                  codebundles_count: { type: integer }
                  support_tags: { type: array, items: { type: string } }
                  pagination:
                    $ref: "#/components/schemas/Pagination"
...
  /github/create-task-request:
    post:
      tags: [GitHub]
      summary: Create codebundle request issue
      operationId: createTaskRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_query, task_description, use_case]
              properties:
                user_query: { type: string }
                task_description: { type: string }
                use_case: { type: string }
                platform: { type: string, default: Generic }
                priority: { type: string, default: medium }
                user_email: { type: string }
      responses:
        "200":
          description: Created issue
          content:
            application/json:
              schema: { type: object }
  /github/issue-template:
    get:
      tags: [GitHub]
      summary: Get issue template
      operationId: issueTemplate
      parameters:
        - name: user_query
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pre-filled template
          content:
            application/json:
              schema: { type: object }

  # =========================================================================
  # Admin - Collections CRUD
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin token (must start with `admin-`)
  parameters:
    CollectionSlug:
      name: collection_slug
      in: path
      required: true
      schema:
        type: string
      description: Collection slug identifier
  schemas:
    Pagination:
      type: object
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
        has_more: { type: boolean }
...
    CollectionWithStats:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, example: AWS Cloud/Custodian CodeCollection }
        slug: { type: string, example: aws-c7n-codecollection }
        description: { type: string }
        git_url: { type: string, format: uri }
        git_ref: { type: string }
        owner: { type: string }
        owner_email: { type: string }
        owner_icon: { type: string }
        is_active: { type: boolean }
        codebundle_count: { type: integer }
        task_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CodeBundle:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string, example: aws-c7n-ebs-health }
        name: { type: string }
        display_name: { type: string, example: AWS EBS Health }
        description: { type: string }
        doc: { type: string }
        author: { type: string }
        support_tags:
          type: array
          items: { type: string }
        discovery_platform: { type: string }
        access_level: { type: string, enum: [read-only, read-write, unknown] }
        task_count: { type: integer }
        sli_count: { type: integer }
        codecollection:
          type: object
          properties:
            id: { type: integer }
            slug: { type: string }
            name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CodeBundleFull:
      allOf:
        - $ref: "#/components/schemas/CodeBundle"
        - type: object
          properties:
            tasks:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  doc: { type: string }
            slis:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  doc: { type: string }
            user_variables:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  description: { type: string }
                  default: { type: string }
                  example: { type: string }
            minimum_iam_requirements:
              type: array
              items: { type: string }
            configuration_type:
              type: object
              properties:
                platform: { type: string }
                has_generation_rules: { type: boolean }
                resource_types:
                  type: array
                  items: { type: string }
                level_of_detail: { type: string }
            ai_enhanced_description: { type: string }
            ai_enhanced_metadata: { type: object }
            readme: { type: string }
...